requirements.txt

# LangChain ecosystem
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0

# Database
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0

langsmith>=0.1.0

aiosqlite>=0.19.0

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
